+++ src/widgets/work-management/ui/WorkManagement.tsx
@@ -1,12 +1,13 @@
 'use client'
 
-import { isOwnerAtom } from '@/shared/atoms/globalAtom'
-import { useAtom } from 'jotai'
+import { useAtom, useAtomValue } from 'jotai'
+import { isOwnerAtom, storeIdAtom } from '@/shared'
 import WorkManagementWorker from './WorkManagementWorker'
 import WorkManagementOwner from './WorkManagementOwner'
 
 export default function WorkManagement() {
   const isOwner = useAtom(isOwnerAtom)
+  const storeId = useAtomValue(storeIdAtom)
 
-  return isOwner ? <WorkManagementOwner /> : <WorkManagementWorker />
+  return isOwner ? <WorkManagementOwner storeId={storeId} /> : <WorkManagementWorker />
 }
